requirements.txt
python-telegram-bot>=21.0
requests>=2.32
apscheduler>=3.10
python-dotenv>=1.0
pytz>=2024.1
